Combine sigsegv signal handlers
This also adds a build step for detecting if we have execinfo.
On Alpine, the previous __GLIBC__ test would fail, but it has
execinfo.
